Info display improved, kill counting added, health potion drops added, player health bug fixed

+++ src/textadventuregame/Enemy.java
@@ -4,7 +4,7 @@ public abstract class Enemy {
     
     int maxHealth;
     int maxAttackDamage;
-    int dropRate;
+    float dropRate;
    
     // Each enemy will have different attack
     public abstract int enemyAttack();
